refactor: Extract decision importing logic into one class
